ingestion_service/requirements.txt
fastapi
uvicorn
apscheduler
qdrant-client
google-genai
requests
python-dotenv
PyPDF2
openai
fastembed
